move `maybe_run_throttled_call` to boto3_connection.py
